+++ src/components/index.js
@@ -64,11 +64,13 @@ export { default as Divider } from './divider';
 export { default as Slider } from './slider';
 export { default as Footer } from './footer';
 export { default as DatePicker } from './datePicker';
+export { default as AvatarUploader } from './avatarUploader';
 export { ThemeContext, themes } from './theme-provider';
 export { 
     convertToNum, 
     compare, 
     phoneMask, 
+    dateMask,
     creditMask,
     creditIssueMask,
     iinMask,